This pull request addresses the issue of inconsistent error handling in the controllers and services, where controllers previously contained pieces of exception handling logic instead.

Removed direct exception handling from controller methods.
Introduced custom exceptions
Updated unit and integration tests to reflect the new error handling flow.

+++ Server/Library/src/main/java/skytales/Library/service/BookService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import org.springframework.web.multipart.MultipartFile;
+import skytales.Library.util.exceptions.*;
 import skytales.Library.util.state_engine.UpdateProducer;
 import skytales.Library.util.state_engine.model.UpdateType;
 import skytales.Library.web.dto.BookData;
@@ -17,9 +18,7 @@ import skytales.Library.util.elasticsearch.service.ElasticSearchService;
 import skytales.Library.model.Book;
 import skytales.Library.repository.BookRepository;
 
-import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Array;
 import java.math.BigDecimal;
 import java.util.*;
 
@@ -45,41 +44,55 @@ public class BookService {
     }
 
     public Book getBookById(UUID bookId) {
-        return bookRepository.findById(bookId).orElse(null);
+        return bookRepository.findById(bookId)
+                .orElseThrow(() -> new BookNotFoundException("No book found with id " + bookId));
     }
 
     public List<Book> getNewestBooks(int year) {
         return bookRepository.findBooksByYearAfter(year)
-                .orElseThrow(() -> new NoSuchElementException("No books found after the year -" + year));
+                .orElseThrow(() -> new BookNotFoundException("No books found after the year " + year));
     }
 
-
     @Transactional
-    public void deleteBooks(ArrayList<String> bookIds) {
+    public void deleteBooks(List<String> bookIds) {
+        if (bookIds == null || bookIds.isEmpty()) {
+            throw new InvalidBookDataException("Book ID list cannot be empty.");
+        }
 
-            bookIds.forEach(bookId -> {
-                try {
+        bookIds.forEach(bookId -> {
+            try {
+                UUID uuid = UUID.fromString(bookId);
+                Book book = bookRepository.findById(uuid)
+                        .orElseThrow(() -> new BookNotFoundException("No book found with id " + bookId));
 
-                    UUID uuid = UUID.fromString(bookId);
-                    Book book = bookRepository.findById(uuid).orElseThrow(() -> new NoSuchElementException("No book found with id " + bookId));
-
-                    updateProducer.sendBookUpdate(UpdateType.REMOVE_BOOK, book);
-                    bookRepository.delete(book);
-                    elasticSearchService.deleteBookFromElasticsearch(bookId);
-
-                } catch (Exception e) {
-                    throw new RuntimeException("Failed to delete book with ID: " + bookId, e);
-                }
-            });
+                updateProducer.sendBookUpdate(UpdateType.REMOVE_BOOK, book);
+                bookRepository.delete(book);
+                elasticSearchService.deleteBookFromElasticsearch(bookId);
+            } catch (Exception e) {
+                throw new BookDeletionException("Failed to delete book with ID: " + bookId, e);
+            }
+        });
     }
 
     @Transactional
     public Book createBook(BookData data, MultipartFile bannerImage, MultipartFile coverImage) throws IOException {
 
         BigDecimal price = new BigDecimal(data.price());
+
         String coverImageUrl = uploadCoverImage(coverImage);
         String bannerImageUrl = uploadCoverImage(bannerImage);
 
+        Book book = buildBook(data, coverImageUrl, bannerImageUrl, price);
+
+        bookRepository.save(book);
+        elasticSearchService.addBookToElasticsearch(book);
+        updateProducer.sendBookUpdate(UpdateType.NEW_BOOK, book);
+
+        log.info("Book created and indexed with title - " + book.getTitle());
+        return book;
+    }
+
+    private static Book buildBook(BookData data, String coverImageUrl, String bannerImageUrl, BigDecimal price) {
         Book book = Book.builder()
                 .title(data.title())
                 .author(data.author())
@@ -91,55 +104,32 @@ public class BookService {
                 .genre(data.genre())
                 .quantity(Integer.parseInt(data.quantity()))
                 .build();
-
-        bookRepository.save(book);
-        elasticSearchService.addBookToElasticsearch(book);
-        updateProducer.sendBookUpdate(UpdateType.NEW_BOOK, book);
-
-        log.info("Book created and indexed with title - " + book.getTitle());
         return book;
     }
 
-//    public String uploadCoverImage(File coverImage) throws IOException {
-//
-//        String fileName = coverImage.getName();
-//        String fileExtension = fileName.substring(fileName.lastIndexOf('.') + 1);
-//
-//        if (!fileExtension.matches("jpg|jpeg|png|webp")) {
-//            throw new IllegalArgumentException("Unsupported file format: " + fileExtension);
-//        }
-//
-//        Map<String, Object> uploadResult = cloudinary.uploader().upload(coverImage,
-//                ObjectUtils.asMap(
-//                        "resource_type", "auto",
-//                        "quality", "auto:good",
-//                        "format", fileExtension,
-//                        "width", 800,
-//                        "crop", "limit"
-//                ));
-//
-//        return (String) uploadResult.get("secure_url");
-//    }
-
-    private String uploadCoverImage(MultipartFile coverImage) throws IOException {
-
+    private String uploadCoverImage(MultipartFile coverImage) {
         String fileName = coverImage.getOriginalFilename();
         String fileExtension = fileName.substring(fileName.lastIndexOf('.') + 1);
 
         if (!fileExtension.matches("jpg|jpeg|png|webp")) {
-            throw new IllegalArgumentException("Unsupported file format: " + fileExtension);
+            throw new InvalidFileFormatException("Unsupported file format: " + fileExtension);
         }
 
-        Map<String, Object> uploadResult = cloudinary.uploader().upload(coverImage.getBytes(),
-                ObjectUtils.asMap(
-                        "resource_type", "auto",
-                        "quality", "auto:good",
-                        "format", fileExtension,
-                        "width", 800,
-                        "crop", "limit"
-                ));
-
-        return (String) uploadResult.get("secure_url");
+        try {
+            Map<String, Object> uploadResult = cloudinary.uploader().upload(
+                    coverImage.getBytes(),
+                    ObjectUtils.asMap(
+                            "resource_type", "auto",
+                            "quality", "auto:good",
+                            "format", fileExtension,
+                            "width", 800,
+                            "crop", "limit"
+                    )
+            );
+            return (String) uploadResult.get("secure_url");
+        } catch (IOException e) {
+            throw new FileUploadException("Failed to upload image: " + fileName, e);
+        }
     }
